Casching problem and Servlet rename
