+++ src/main/java/com/tlab/mvc/schedule/controller/ScheduleInsertServletByAjax.java
@@ -1,0 +1,25 @@
+package com.tlab.mvc.schedule.controller;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * Servlet implementation class ScheduleInsertServlet
+ */
+@WebServlet("/member/InsertSchedule")
+public class ScheduleInsertServletByAjax extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+
+	/**
+	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
+	 */
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		
+		
+	}
+
+}
